007_mcp/wikipedia-chat-app/mcp-server/requirements.txt
fastmcp
httpx
anthropic
